remove HMapPTableProjector into _projectors module
