+++ pymatviz/ptable/_projector.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import math
 from collections.abc import Sequence
-from typing import TYPE_CHECKING, Literal
+from typing import TYPE_CHECKING, Literal, NamedTuple
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt
@@ -21,7 +21,7 @@ from pymatviz.ptable._process_data import (
     SupportedDataType,
     SupportedValueType,
 )
-from pymatviz.utils import df_ptable
+from pymatviz.utils import df_ptable, pick_bw_for_contrast
 
 
 if TYPE_CHECKING:
@@ -29,6 +29,32 @@ if TYPE_CHECKING:
 
     import pandas as pd
     from matplotlib.typing import ColorType
+    from numpy.typing import NDArray
+
+    from pymatviz.ptable.ptable_matplotlib import ElemStr
+
+
+class TileValueColor(NamedTuple):
+    """Value and colors for heatmap tiles.
+
+    value (str): The value to display.
+    text_color (ColorType): Color for both symbol and value.
+    tile_color (ColorType): Color for the tile.
+    """
+
+    value: str | float
+    text_color: ColorType
+    tile_color: ColorType
+
+
+class OverwriteTileValueColor(TileValueColor):
+    """Overwrite Value and colors for heatmap tiles,
+    where None is used for not overwriting.
+    """
+
+    value: str | float | None  # type: ignore[assignment]
+    text_color: ColorType | None
+    tile_color: ColorType | None
 
 
 class PTableProjector:
@@ -89,7 +115,7 @@ class PTableProjector:
         self.elem_colors = elem_colors  # type: ignore[assignment]
 
         # Preprocess data
-        self.log = log
+        self.log: bool = log
         self.data: pd.DataFrame = data
 
         # Remove excluded element from internal data to avoid metadata pollution
@@ -641,3 +667,222 @@ class ChildPlotters:
         # Hide the right, left and top spines
         ax.axis("on")
         ax.spines[["right", "top", "left"]].set_visible(False)
+
+
+class HMapPTableProjector(PTableProjector):
+    """With more heatmap-specific functionalities."""
+
+    def generate_tile_value_colors(
+        self,
+        *,
+        text_colors: Literal["AUTO"] | ColorType | dict[ElemStr, ColorType] = "AUTO",
+        overwrite_tiles: dict[ElemStr, TileValueColor],
+        inf_color: ColorType,
+        nan_color: ColorType = "lightgrey",
+        excluded_tile_color: ColorType = "white",
+    ) -> dict[ElemStr, TileValueColor]:
+        """Generate value and colors for element tiles.
+
+        Args:
+            text_colors: Colors for element symbols and values.
+                - "AUTO": Auto pick "black" or "white" based on the contrast
+                    of tile color for each element.
+                - ColorType: Use the same ColorType for each element.
+                - dict[ElemStr, ColorType]: Element to color mapping.
+            overwrite_tiles (dict[ElemStr, TileValueColor]): Final
+                entried to overwrite tile value and colors. Note this
+                would overwrite everything, include exclusion and anomalies.
+            inf_color (ColorType): Color for infinities.
+            nan_color (ColorType): Color for missing value (NaN).
+            excluded_tile_color (ColorType): Color for excluded element tile.
+
+        Returns:
+            dict[ElemStr, TileValueColor]: Element to value-color mapping.
+        """
+        # Build TileValueColor for NaN (or absent elements) and infinity
+        inf_tile = TileValueColor("∞", pick_bw_for_contrast(inf_color), inf_color)
+        nan_tile = TileValueColor("-", pick_bw_for_contrast(nan_color), nan_color)
+
+        tile_entries: dict[ElemStr, TileValueColor] = {}
+
+        for element in Element:
+            symbol = element.symbol
+
+            # Skip excluded elements
+            if symbol in self.exclude_elements:
+                tile_entries[symbol] = TileValueColor(
+                    "excl.",
+                    pick_bw_for_contrast(excluded_tile_color),
+                    excluded_tile_color,
+                )
+                continue
+
+            # Handle NaN/infinity colors and values
+            if self.anomalies != "NA" and symbol in self.anomalies:
+                if "inf" in self.anomalies[symbol]:
+                    tile_entries[symbol] = inf_tile
+
+                # Note: For heatmap plotter, ideally there should not
+                # be NaN in the value, but this might happen if
+                # NaNs are not dropped
+                else:
+                    tile_entries[symbol] = nan_tile
+
+                continue
+
+            # Try to get data from DataFrame
+            try:
+                values: NDArray = self.data.loc[symbol, Key.heat_val]
+
+                if len(values) != 1:
+                    raise ValueError(f"Data for {symbol} should be length 1.")
+
+                value = float(values[0])
+
+                # Generate tile color and text color
+                if self.cmap is None:
+                    raise ValueError("Cannot generate tile color without colormap.")
+                tile_color = self.cmap(self.norm(value))
+
+                if text_colors == "AUTO":
+                    text_color: str = pick_bw_for_contrast(tile_color)
+                elif isinstance(text_colors, dict):
+                    text_color = text_colors.get(symbol, "black")
+                else:
+                    text_color = text_colors
+
+                tile_entries[symbol] = TileValueColor(value, text_color, tile_color)
+
+            # For element absent from data, use "-"
+            except KeyError:
+                tile_entries[symbol] = nan_tile
+
+        # Apply overwrite colors
+        for symbol, ow_tile_entry in overwrite_tiles.items():
+            tile_entries[symbol] = TileValueColor(
+                ow_tile_entry.value or tile_entries[symbol].value,
+                ow_tile_entry.text_color or tile_entries[symbol].text_color,
+                ow_tile_entry.tile_color or tile_entries[symbol].tile_color,
+            )
+
+        return tile_entries
+
+    def add_heatmap_tiles(
+        self,
+        tile_entries: dict[str, TileValueColor],
+        *,
+        f_block_voffset: float = 0,  # noqa: ARG002 TODO: fix this
+        symbol_pos: tuple[float, float],
+        symbol_kwargs: dict[str, Any] | None = None,
+        sci_notation: bool,
+        value_show_mode: Literal["value", "fraction", "percent", "off"],
+        value_fmt: str,
+        value_pos: tuple[float, float],
+        value_kwargs: dict[str, Any] | None = None,
+        ax_kwargs: dict[str, Any] | None = None,
+    ) -> None:
+        """Add heatmap tiles (element symbol, tile color and optional value)
+        to the periodic table grid.
+
+        Args:
+            tile_entries (dict[str, TileValueColor]): Entries for each tile.
+            f_block_voffset (float): The vertical offset of f-block elements.
+            symbol_pos (tuple[float, float]): Position of element symbols
+                relative to the lower left corner of each tile.
+                Defaults to (0.5, 0.5). (1, 1) is the upper right corner.
+            symbol_kwargs (dict): Keyword arguments passed to plt.text() for
+                element symbols. Defaults to None.
+            sci_notation (bool): Whether to use scientific notation for values and
+                colorbar tick labels.
+            value_show_mode (str): The values display mode:
+                - "off": Hide values.
+                - "value": Display values as is.
+                - "fraction": As a fraction of the total (0.10).
+                - "percent": As a percentage of the total (10%).
+                "fraction" and "percent" can be used to make the colors in
+                    different plots comparable.
+            value_pos (tuple[float, float]): The position of values inside the tile.
+            value_fmt (str | "AUTO"): f-string format for values. Defaults to ".1%"
+                (1 decimal place) if values_show_mode is "percent", else ".3g".
+            value_kwargs (dict): Keyword arguments passed to plt.text() for
+                values. Defaults to None.
+            ax_kwargs (dict): Keyword arguments to pass to ax.set().
+        """
+        # Update kwargs
+        ax_kwargs = ax_kwargs or {}
+        symbol_kwargs = symbol_kwargs or {"fontsize": 16, "fontweight": "bold"}
+        value_kwargs = value_kwargs or {}
+        if sci_notation:
+            value_kwargs.setdefault("fontsize", 10)
+        else:
+            value_kwargs.setdefault("fontsize", 12)
+
+        for element in Element:
+            # Hide f-block
+            if self.hide_f_block and (element.is_lanthanoid or element.is_actinoid):
+                continue
+
+            # Get axis index by element symbol
+            symbol: str = element.symbol
+            row, column = df_ptable.loc[symbol, ["row", "column"]]
+            ax: plt.Axes = self.axes[row - 1][column - 1]
+
+            # Unpack tile entry
+            value, text_color, tile_color = tile_entries[symbol]
+
+            # Skip element without data if "hide"
+            if value == "-" and self.on_empty == "hide":
+                continue
+
+            # Add tile background color
+            ax.pie(
+                [1.0],  # a single pie plot for cropping
+                colors=[tile_color],
+                wedgeprops={"clip_on": True},
+            )
+            rect = Rectangle(
+                xy=(-0.5, -0.5), width=1, height=1, fc="none", ec="grey", lw=2
+            )
+            ax.add_patch(rect)
+
+            ax.set_xlim(-0.5, 0.5)
+            ax.set_ylim(-0.5, 0.5)
+
+            # Add element symbol
+            ax.text(
+                *symbol_pos,
+                symbol,
+                color=text_color,
+                ha="center",
+                va="center",
+                transform=ax.transAxes,
+                **symbol_kwargs,
+            )
+
+            # (Upon request) Show value
+            if value_show_mode == "off":
+                continue
+
+            # Format value
+            try:
+                value = f"{value:{value_fmt}}"
+            except ValueError:
+                value = str(value)
+
+            # Simplify scientific notation, e.g. 1e-01 to 1e-1
+            if sci_notation and ("e-0" in value or "e+0" in value):
+                value = value.replace("e-0", "e-").replace("e+0", "e+")
+
+            ax.text(
+                *value_pos,
+                value,
+                color=text_color,
+                ha="center",
+                va="center",
+                transform=ax.transAxes,
+                **value_kwargs,
+            )
+
+            # Pass axis kwargs
+            if ax_kwargs:
+                ax.set(**ax_kwargs)
